More tweaks to codchess

+++ codchess/include/codchess/perft.hpp
@@ -7,5 +7,5 @@ namespace codchess::perft {
 /// @param board The board to move generation on
 /// @param depth The depth to search to
 /// @return The number of nodes found
-std::uint64_t Perft(Board &board, std::uint64_t depth) noexcept;
+std::uint64_t Perft(Board &board, std::uint32_t depth) noexcept;
 }   // namespace codchess::perft
